# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NHM Product Dat Xe Backend API",
        "description": "API documentation for NHM Product Dat Xe Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/driver/status": {
            "put": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-31: Bật/Tắt trạng thái hoạt động (Go Online/Offline)",
                "description": "Tài xế có thể bật/tắt trạng thái bất cứ lúc nào. Nếu tắt Offline khi đang có chuyến, tài xế vẫn hoàn thành chuyến cũ nhưng sẽ không được gán thêm chuyến mới.",
                "operationId": "800c64bb5a49044aae8e087ebe3ec0ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_online"
                                ],
                                "properties": {
                                    "is_online": {
                                        "description": "True để Online, False để Offline",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "current_lat": {
                                        "description": "Bắt buộc nếu is_online = true",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.776889
                                    },
                                    "current_lng": {
                                        "description": "Bắt buộc nếu is_online = true",
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.700806
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật trạng thái thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "403": {
                        "description": "Tài khoản chưa được duyệt, bị khóa hoặc đang chờ (Cooldown)"
                    },
                    "422": {
                        "description": "Lỗi xử lý nghiệp vụ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/ride/{rideId}/accept": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-32: Chấp nhận đơn hàng/chuyến xe",
                "operationId": "519c3194f7234a35cfae24430a69939f",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_lat",
                                    "current_lng"
                                ],
                                "properties": {
                                    "current_lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.776889
                                    },
                                    "current_lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.700806
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nhận đơn thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "403": {
                        "description": "Không đủ điều kiện nhận đơn"
                    },
                    "422": {
                        "description": "Đơn không còn khả dụng hoặc driver đang bận"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/ride/{rideId}/reject": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-33: Từ chối đơn hàng (Trước khi nhận)",
                "operationId": "fc32e58825231b8b5b786b9ea5ec8189",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Từ chối thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    },
                    "422": {
                        "description": "Đơn không ở trạng thái chờ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/ride/{rideId}/cancel": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-33: Hủy chuyến đi (Sau khi đã nhận)",
                "operationId": "e0e1a8bc462e227f5491d4cdc940b1a4",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason_id"
                                ],
                                "properties": {
                                    "reason_id": {
                                        "description": "ID lý do hủy chuyến. 1: Khách không ra (Customer No Show), 2: Xe hỏng (Vehicle Broken), 3: Đặt sai điểm (Wrong Location), 4: Khác (Other)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "current_lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.776889
                                    },
                                    "current_lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.700806
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hủy thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "403": {
                        "description": "Không có quyền hủy hoặc bị phạt"
                    },
                    "422": {
                        "description": "Trạng thái đơn không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/ride/{rideId}/start": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-35: Bắt đầu thực hiện chuyến đi (Start Trip)",
                "operationId": "ec53fb0b8e81b9348b463667caf839d9",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.776889
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.700806
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bắt đầu thành công"
                    },
                    "403": {
                        "description": "Không có quyền sở hữu chuyến hoặc chưa đủ gần"
                    },
                    "422": {
                        "description": "Trạng thái đơn không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/ride/{rideId}/complete": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-40: Hoàn thành chuyến đi (Complete Trip)",
                "operationId": "2d78e2e031d0efb77bacc9bb0375b91f",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.776889
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.700806
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hoàn thành chuyến đi thành công"
                    },
                    "403": {
                        "description": "Không có quyền sở hữu chuyến hoặc chưa đủ gần điểm đến"
                    },
                    "422": {
                        "description": "Trạng thái đơn không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/ride/{rideId}/arrived": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "UC-36: Thông báo đã đến điểm đón (Arrived)",
                "operationId": "579713bda37acdf54c1f165d30daaaaa",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.776889
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.700806
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thông báo thành công"
                    },
                    "403": {
                        "description": "Không có quyền sở hữu chuyến hoặc chưa đủ gần"
                    },
                    "422": {
                        "description": "Trạng thái đơn không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}